requirements.txt
# Core dependencies
langchain
langchain-core
langchain-openai
langchain-community
langchain-pinecone
langgraph
pinecone-client
streamlit


# Utilities
python-dotenv
pyyaml
Pillow  # For image processing

# Observability
langfuse
langsmith
loguru

# Data processing
unstructured
tiktoken
markdown

# Development tools
pytest
black 